# openapi/model_game_container_configuration.go
/*
 * GameBase Communication API
 *
 * This is the REST API used as an communication layer between backend and frontend.
 *
 * API version: 2.0.0
 * Contact: gamebase@gahr.dev
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type GameContainerConfiguration struct {
	Details GameContainerConfigurationDetails `json:"details"`

	Resources GameContainerConfigurationResources `json:"resources,omitempty"`
}