# openapi/model_game_container_deployment.go
/*
 * GameBase Communication API
 *
 * This is the REST API used as an communication layer between backend and frontend.
 *
 * API version: 2.0.0
 * Contact: gamebase@gahr.dev
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type GameContainerDeployment struct {

	// Template path of backend directory that is going to be used for game container creation
	TemplatePath string `json:"templatePath"`
}