# openapi/model_game_container_status.go
/*
 * GameBase Communication API
 *
 * This is the REST API used as an communication layer between backend and frontend.
 *
 * API version: 2.0.0
 * Contact: gamebase@gahr.dev
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type GameContainerStatus struct {

	// ID of game server container
	Id string `json:"id,omitempty"`

	Status Status `json:"status,omitempty"`

	Configuration GameContainerConfiguration `json:"configuration,omitempty"`

	// Dynamic details by game server monitoring agent
	GameServerDetails map[string]string `json:"gameServerDetails,omitempty"`
}