# openapi/model_restart_behavior.go
/*
 * GameBase Communication API
 *
 * This is the REST API used as an communication layer between backend and frontend.
 *
 * API version: 2.0.0
 * Contact: gamebase@gahr.dev
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type RestartBehavior string

// List of RestartBehavior
const (
	NONE           RestartBehavior = "none"
	UNLESS_STOPPED RestartBehavior = "unless-stopped"
	ON_FAILURE     RestartBehavior = "on-failure"
	ALWAYS         RestartBehavior = "always"
)