# openapi/model_status.go
/*
 * GameBase Communication API
 *
 * This is the REST API used as an communication layer between backend and frontend.
 *
 * API version: 2.0.0
 * Contact: gamebase@gahr.dev
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Status string

// List of Status
const (
	UNKNOWN    Status = "UNKNOWN"
	RUNNING    Status = "RUNNING"
	STOPPED    Status = "STOPPED"
	RESTARTING Status = "RESTARTING"
	ERROR      Status = "ERROR"
)