# openapi/model_user.go
/*
 * GameBase Communication API
 *
 * This is the REST API used as an communication layer between backend and frontend.
 *
 * API version: 2.0.0
 * Contact: gamebase@gahr.dev
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type User struct {

	// Email address of the user
	Email string `json:"email"`

	// The full name of the user
	FullName string `json:"fullName"`

	// The JWT of the users session
	Token string `json:"token"`
}