# openapi/model_user_login.go
/*
 * GameBase Communication API
 *
 * This is the REST API used as an communication layer between backend and frontend.
 *
 * API version: 2.0.0
 * Contact: gamebase@gahr.dev
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type UserLogin struct {

	// Email address of the user
	Email string `json:"email"`

	// The password of the user
	Password string `json:"password"`
}